# api/openapi.yaml
openapi: 3.0.3
info:
  title: Experimental Game Server API
  version: 0.4.0
  description: |
    Contract-first REST API for an experimental Go game web server.

    This API is intentionally small and churn-friendly. Treat this file as the
    public contract for player-written clients. Server DTOs and handler stubs are
    generated from this file with oapi-codegen.

    ## Game management

    Everything up to (but not including) the game engine is implemented: creating
    games, managing their lifecycle, and managing rosters. The engine — turns and
    orders (`listTurns`, `getTurn`, `validateOrders`, `submitOrders`,
    `getOrderSubmission`) — is still stubbed and answers `501 Not Implemented`.

    Roles are global-admin, GM, and player. An **admin** creates games and has
    god-mode over every game; admins never hold a membership. A **GM** or
    **player** is a membership row on one game (a GM has elevated rights there).

    A game's status advances along a forward-only chain, skips allowed:
    `draft → recruiting → active → paused → complete → archived`. The only
    backward move is `paused → active` (un-pause), which is admin-only, as is
    moving a game out of `archived`; an archived game is otherwise frozen.

    ### Typical create → recruit → activate workflow

    1. `createGame` (admin) — a new game starts in `draft`.
    2. `addGameMember` with `isGm: true` (admin) — assign the first GM.
...
    5. `updateGame` `status: active` (GM or admin) — begin play.

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Exchange a refresh token for a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refresh succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /me/games:
    get:
      tags: [Me]
      operationId: listMyGames
      summary: List the games the authenticated user participates in
      description: |
        Returns the games the caller is currently a member of, each with the
        caller's handle in that game and whether they are its game master.
        Scoped to the caller (self-service); no admin role is required. This is
        a read-only projection of the caller's memberships and runs no game
        logic.
      responses:
        '200':
          description: The caller's games
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMyGamesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/password:
    post:
      tags: [Me]
      operationId: changeMyPassword
      summary: Change the authenticated user's password
      description: |
        Self-service password change for the logged-in account. The caller must
        supply their current password, which is verified against the stored hash
        before the new password is applied. On success the new password is hashed
        and stored, and the caller's other sessions are revoked so a stolen
        session cannot outlive the password it was created under. No admin role is
        required; the account state is re-read fresh (an account deactivated since
        the token was issued cannot change its password).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/shutdown:
    post:
      tags: [Admin]
      operationId: shutdownServer
      summary: Gracefully shut down the server (development only)
      description: |
        Development only. Triggers a graceful shutdown: the server stops
        accepting new connections and drains in-flight requests (including this
        one) before the process exits. Enabled only when the server was started
        with --development; otherwise it responds 404 as if the route did not
        exist. Admin only.
      responses:
        '202':
          description: Shutdown initiated; the server will drain and stop
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/refresh-tokens/purge:
    post:
      tags: [Admin]
      operationId: purgeRefreshTokens
      summary: Purge expired refresh tokens
      description: |
        Admin only. Deletes refresh-token records that have already expired,
        reclaiming space in the refresh_tokens table. Expired tokens no longer
        authenticate — they fail signature/expiry verification before any store
        lookup — so removing them affects neither active sessions nor
        theft/reuse detection. Revoked-but-unexpired tokens are intentionally
        retained, since presenting one is the reuse signal that revokes a
        session family. Returns the number of records removed. A background
        reaper performs the same purge periodically; this endpoint runs it on
        demand.
      responses:
        '200':
          description: Purge completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurgeRefreshTokensResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /games:
    get:
      tags: [Games]
      operationId: listGames
      summary: List games visible to the current user
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/GameStatus'
      responses:
        '200':
          description: Visible games
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListGamesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Games]
      operationId: createGame
      summary: Create a game
      description: Requires an admin role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
            examples:
              basic:
                value:
                  code: ALPHA
                  name: Alpha Campaign
                  description: The first playtest game.
      responses:
        '201':
          description: Game created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /games/{gameId}:
    parameters:
      - $ref: '#/components/parameters/GameId'
    get:
      tags: [Games]
      operationId: getGame
      summary: Get a game by id
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Games]
      operationId: updateGame
      summary: Update a game (status, metadata, visibility)
      description: |
        Applies a partial update to a game. Each field has its own rule from the
        action matrix:
          - status: advances along draft → recruiting → active → paused →
            complete → archived (forward-only, skips allowed) for an active GM or
            an admin. Backward moves are rejected with a 409 except paused → active
            (un-pause) and moving out of archived, both admin-only.
          - name / description: an active GM or an admin; rejected once archived.
          - isActive: admin only; hides the game from non-admin listings,
            orthogonal to status. Frozen while archived.
        An archived game is otherwise frozen: the only accepted change is an admin
        moving its status to another state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGameRequest'
            examples:
              openRecruiting:
                summary: Open the game for players (draft → recruiting)
                value:
                  status: recruiting
              activate:
                summary: Begin play (recruiting → active)
                value:
                  status: active
              rename:
                summary: Edit metadata
                value:
                  name: Alpha Campaign
                  description: An updated blurb.
              hide:
                summary: Admin hard-hide
                value:
                  isActive: false
      responses:
        '200':
          description: The updated game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /games/{gameId}/turns:
    parameters:
      - $ref: '#/components/parameters/GameId'
    get:
      tags: [Turns]
      operationId: listTurns
      summary: List turns for a game
      responses:
        '200':
          description: Turns for the game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTurnsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /games/{gameId}/turns/{turnId}:
    parameters:
      - $ref: '#/components/parameters/GameId'
      - $ref: '#/components/parameters/TurnId'
    get:
      tags: [Turns]
      operationId: getTurn
      summary: Get a turn by id
      responses:
        '200':
          description: Turn details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Turn'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /games/{gameId}/turns/{turnId}/orders:validate:
    parameters:
      - $ref: '#/components/parameters/GameId'
      - $ref: '#/components/parameters/TurnId'
    post:
      tags: [Orders]
      operationId: validateOrders
      summary: Validate an order set without submitting it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrdersRequest'
            examples:
              basic:
                value:
                  factionId: 7
                  orders: |
                    UNIT 101 MOVE N
                    UNIT 102 SCOUT E
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderValidationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
...
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresInSeconds]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresInSeconds:
          type: integer
          minimum: 1
          example: 900
...
    GameStatus:
      type: string
      enum: [draft, recruiting, active, paused, complete, archived]
...